docs+style(streaming): slice-3 doc.go, drop PERF, collision test, lint
- doc.go (resolved in the rebase): scope the file map to the complete
  daemon, add window.go (geometry) + an Index group (txindex.go), drop the
  forbidden design-docs/full-history-implementation-status.md reference,
  prefer 'catalog' over 'meta-store'.
- Remove PERF.md + perf_test.go: the bench-format-alignment material
  belongs with the bench harness, not the daemon PR.
- Add TestBuildTxhashIndex_SameWindowKeyCollisionFailsLoud: a same-window
  16-byte-prefix collision must fail loudly with streamhash.ErrDuplicateKey
  (issue #814 acceptance), never silently drop — previously uncovered.
- golangci-lint (this slice's own new findings): gci/misspell/modernize/
  unconvert via --fix; //nolint:cyclop on buildTxhashIndex; revive unused
  pendingArtifacts cfg -> _; lll wrap; //nolint:unparam on the general test
  helpers; //nolint:funlen,cyclop,maintidx on the lookup E2E.

+++ cmd/stellar-rpc/internal/fullhistory/streaming/daemon_test.go
@@ -137,7 +137,7 @@ func TestRunDaemon_LoadValidateWireStartCleanShutdown(t *testing.T) {
 	cpi, cpiPinned, err := cat.ChunksPerTxhashIndex()
 	require.NoError(t, err)
 	require.True(t, cpiPinned)
-	assert.Equal(t, uint32(DefaultChunksPerTxhashIndex), cpi)
+	assert.Equal(t, DefaultChunksPerTxhashIndex, cpi)
 }
 
 // Storage-path overrides must be HONORED by the data path, not just locked. The
